backend/requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
boto3==1.34.145
botocore==1.34.145
cachetools==5.4.0
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.2.16
django-cleanup==8.0.0
django-cors-headers==4.3.1
django-crontab==0.7.1
django-dbbackup==4.1.0
django-debug-toolbar==4.4.6
django-filter==23.3
django-storages==1.14.4
djangorestframework==3.14.0
drf-extra-fields==3.7.0
filetype==1.2.0
google-api-core==2.19.1
google-api-python-client==2.139.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.63.2
gunicorn==22.0.0
httplib2==0.22.0
idna==3.7
jmespath==1.0.1
oauthlib==3.2.2
packaging==24.0
Pillow==10.1.0
proto-plus==1.24.0
protobuf==5.27.3
psycopg2==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2023.3.post1
redis==5.0.7
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
s3transfer==0.10.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.19
